Split window into two modules
